pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pretty-cool-events"
version = "1.0.0"
description = "Illumio PCE event monitoring and notification system"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "Flask>=3.0",
    "httpx>=0.27",
    "Jinja2>=3.1",
    "pydantic>=2.5",
    "PyYAML>=6.0",
    "rich>=13.0",
    "slack_sdk>=3.11",
    "boto3>=1.28",
    "jira>=3.5",
    "pdpyras>=5.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
    "types-requests",
]

[project.scripts]
pce-events = "pretty_cool_events.cli:cli"

[tool.setuptools.packages.find]
include = ["pretty_cool_events*"]

[tool.setuptools.package-data]
pretty_cool_events = ["data/*.yaml", "templates/*", "web/templates/*", "web/static/*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
